[SPARK-22111][MLLIB] OnlineLDAOptimizer should filter out empty documents beforehand
batchSize should be Long.

+++ mllib/src/main/scala/org/apache/spark/mllib/clustering/LDAOptimizer.scala
@@ -514,7 +514,7 @@ final class OnlineLDAOptimizer extends LDAOptimizer with Logging {
   /**
    * Update lambda based on the batch submitted. batchSize can be different for each iteration.
    */
-  private def updateLambda(stat: BDM[Double], batchSize: Double): Unit = {
+  private def updateLambda(stat: BDM[Double], batchSize: Long): Unit = {
     // weight of the mini-batch.
     val weight = rho()
